restore supply's location and splite deduct fees
