+++ app/ante/account.go
@@ -2,7 +2,7 @@ package ante
 
 import (
 	"bytes"
-	exported2 "github.com/okex/exchain/libs/cosmos-sdk/x/supply/exported"
+	exported2 "github.com/okex/exchain/libs/cosmos-sdk/x/auth/types"
 	"math/big"
 
 	"github.com/okex/exchain/libs/cosmos-sdk/x/auth/exported"
