+++ libs/cosmos-sdk/x/auth/ante/fee.go
@@ -2,7 +2,7 @@ package ante
 
 import (
 	"fmt"
-	exported2 "github.com/okex/exchain/libs/cosmos-sdk/x/supply/exported"
+	"log"
 
 	sdk "github.com/okex/exchain/libs/cosmos-sdk/types"
 	"github.com/okex/exchain/libs/cosmos-sdk/x/auth/exported"
@@ -75,10 +75,10 @@ func (mfd MempoolFeeDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate b
 // CONTRACT: Tx must implement FeeTx interface to use DeductFeeDecorator
 type DeductFeeDecorator struct {
 	ak           keeper.AccountKeeper
-	supplyKeeper exported2.SupplyKeeper
+	supplyKeeper types.SupplyKeeper
 }
 
-func NewDeductFeeDecorator(ak keeper.AccountKeeper, sk exported2.SupplyKeeper) DeductFeeDecorator {
+func NewDeductFeeDecorator(ak keeper.AccountKeeper, sk types.SupplyKeeper) DeductFeeDecorator {
 	return DeductFeeDecorator{
 		ak:           ak,
 		supplyKeeper: sk,
@@ -102,6 +102,10 @@ func (dfd DeductFeeDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bo
 		return ctx, sdkerrors.Wrapf(sdkerrors.ErrUnknownAddress, "fee payer address: %s does not exist", feePayer)
 	}
 
+	if ctx.IsDeliver() {
+		txtype := feeTx.GetType()
+		log.Println(txtype)
+	}
 	// deduct the fees
 	if !feeTx.GetFee().IsZero() {
 		err = DeductFees(dfd.supplyKeeper, ctx, feePayerAcc, feeTx.GetFee())
@@ -117,7 +121,28 @@ func (dfd DeductFeeDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bo
 //
 // NOTE: We could use the BankKeeper (in addition to the AccountKeeper, because
 // the BankKeeper doesn't give us accounts), but it seems easier to do this.
-func DeductFees(supplyKeeper exported2.SupplyKeeper, ctx sdk.Context, acc exported.Account, fees sdk.Coins) error {
+func DeductFees(supplyKeeper types.SupplyKeeper, ctx sdk.Context, acc exported.Account, fees sdk.Coins) error {
+	err := checkAccountAndDeductFee(supplyKeeper, ctx, acc, fees)
+	if err != nil {
+		return err
+	}
+
+	return supplyKeeper.AddCoinsToFeeCollector(ctx, fees)
+}
+
+func DeductEvmFees(supplyKeeper types.SupplyKeeper, ctx sdk.Context, acc exported.Account, fees sdk.Coins) error {
+	err := checkAccountAndDeductFee(supplyKeeper, ctx, acc, fees)
+	if err != nil {
+		return err
+	}
+	if ctx.IsDeliver() {
+		supplyKeeper.AddFee(fees)
+	}
+
+	return nil
+}
+
+func checkAccountAndDeductFee(supplyKeeper types.SupplyKeeper, ctx sdk.Context, acc exported.Account, fees sdk.Coins) error {
 	blockTime := ctx.BlockTime()
 	coins := acc.GetCoins()
 
@@ -140,9 +165,6 @@ func DeductFees(supplyKeeper exported2.SupplyKeeper, ctx sdk.Context, acc export
 			"insufficient funds to pay for fees; %s < %s", spendableCoins, fees)
 	}
 
-	if ctx.IsDeliver() {
-		supplyKeeper.AddFee(fees)
-	}
 	err := supplyKeeper.SubtractCoins(ctx, acc.GetAddress(), fees)
 	if err != nil {
 		return sdkerrors.Wrapf(sdkerrors.ErrInsufficientFunds, err.Error())
